create restaurant screen with card first steps, and add styled components
