+++ src/features/restaurants/componens/restaurants-info-card.component.js
@@ -1,0 +1,33 @@
+import React from 'react';
+import styled from 'styled-components/native';
+import { StyleSheet } from 'react-native';
+import { Card } from 'react-native-paper';
+
+const Title = styled.Text`
+  padding: 16px;
+`;
+
+export const RestaurantInfoCard = ({ restaurant = {} }) => {
+  const {
+    name = 'Some Restaurant',
+    icon,
+    photos = [
+      'https://www.foodiesfeed.com/wp-content/uploads/2019/06/top-view-for-box-of-2-burgers-home-made-600x899.jpg',
+    ],
+    address = '100 some random street',
+    isOpenNow = true,
+    rating = 4,
+    isClosedTemporarily,
+  } = restaurant;
+  return (
+    <Card>
+      <Card.Cover key={name} source={{ uri: photos[0] }} style={styles.cover} />
+      <Title>{name}</Title>
+    </Card>
+  );
+};
+
+const styles = StyleSheet.create({
+  card: { backgroundColor: 'white' },
+  cover: { padding: 20, backgroundColor: 'white' },
+});
